refactor the schema settings related code into custom hook
