+++ app/packages/state/src/hooks/index.ts
@@ -48,3 +48,4 @@ export { default as useToEvaluationPatches } from "./useToEvaluationPatches";
 export { default as useTooltip } from "./useTooltip";
 export { default as useToPatches } from "./useToPatches";
 export { default as useUpdateSamples } from "./useUpdateSamples";
+export { default as useSchemaSettings } from "./useSchemaSettings";
